requirements.txt
pystache
PyYAML
setuptools
graphviz